+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cepweb",
-  "version": "1.0.2",
+  "version": "1.0.4",
   "description": "Carpet Cutters Installer App",
   "productName": "CEP Web",
   "author": "XLautomation <scott@xlautomation.com.au>",
