+++ src/Services/CheckReset.aspx
@@ -13,16 +13,13 @@
     '   Request parameter: none
     '   Response: session email and reset key, if set
     ' 
-    Dim strRequestEmail As String
-    strRequestEmail = Request.Headers("UserEmail")
     Dim strUserEmail As String
     strUserEmail = Session("Email")
-	PutLog("CEP app", "Installer", strRequestEmail, "Debug", "CheckReset", "Started")
+ 	PutLog("CEP app", "Installer", strUserEmail, "Debug", "CheckReset", "Started")
     Dim strResetKey As String
     strResetKey = Session("ResetKey")
 	Session("ResetKey") = ""
 	
     Response.ContentType="application/json; charset=utf-8" 
 	Response.Write("{""email"": """ & strUserEmail & """, ""resetKey"": """ & strResetKey & """}")
-
 %>
