+++ src/layouts/Auth.vue
@@ -21,43 +21,8 @@
 
 <script>
 
-//import globalData from 'src/components/GlobalData.js'
-import { SessionStorage } from 'quasar';
-
 export default {
-  mounted () {
-    //
-    //    Look for password reset data stored in ASP session
-    //    - if present, navigate to reset password page
-    //    - otherwise, navigate to login page
-    //
-    const xhttp = new XMLHttpRequest();
-    xhttp.onload = () => {
-      var resultObj = JSON.parse(xhttp.response);
-      //console.log(resultObj);
-      SessionStorage.set('userEmail',resultObj.email);      //  used by Login/ChangePW page
-      if (resultObj.resetKey == '') {   //  Normal login
-        this.$router.push('Login');
-      }
-      else {   //  Reset password link has been used
-        SessionStorage.set('resetKey',resultObj.resetKey);     //  used by ChangePW page
-        this.$router.push('ResetPW');
-      }
-    }
-    xhttp.open('GET', 'Services/CheckReset.aspx', true);
-    xhttp.send();
-
-  }
 
 }
 
 </script>
-<style>
-
-  .bg-image {
-    background-image: url(https://cdn.quasar.dev/img/mountains.jpg);
-    background-repeat: no-repeat;
-    background-size: contain;
-  }
-
-</style>
