+++ src/pages/ResetPW.vue
@@ -66,6 +66,7 @@ import { SessionStorage } from 'quasar';
 export default {
   setup() {
     const $q = useQuasar();
+    
     return {
 
       showUploadNote (text, colour) {
@@ -81,17 +82,41 @@ export default {
   data () {
 
     return {
-      userEmail: SessionStorage.getItem('userEmail'),
-      resetKey: SessionStorage.getItem('resetKey'),
+      userEmail: '',
+      resetKey: '',
       password: '',
       password2: '',
       resultMsg: '',
       resultType: '',
       resultTitle: '',
-      userMessage: '',
+      userMessage: ref(false),
       isPwd: ref(true),
       isPwd2: ref(true),
     }
+  },
+
+  mounted () {
+    //
+    //    Look for password reset data stored in ASP session
+    //    - if present, continue
+    //    - otherwise, navigate to login page
+    //
+    const xhttp = new XMLHttpRequest();
+    xhttp.onload = () => {
+      var resultObj = JSON.parse(xhttp.response);
+      //console.log(resultObj);
+      SessionStorage.set('userEmail',resultObj.email);
+      if (resultObj.resetKey == '') {   //  Direct to normal login - no errors returned
+        this.$router.push('Login');
+      }
+      else {   //  Reset password email link has been used
+        this.resetKey = resultObj.resetKey; 
+        this.userEmail = resultObj.email;
+      }
+    }
+    xhttp.open('GET', 'Services/CheckReset.aspx', true);
+    xhttp.send();
+
   },
 
   methods: {
